ml-service/requirements.txt
flask==3.1.2
flask-cors==6.0.2
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.1.3
joblib==1.4.2
gunicorn==23.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
optuna==3.6.1
shap==0.45.1
xgboost==2.0.3
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation-flask==0.46b0
opentelemetry-exporter-otlp==1.25.0
